Keep provider state in providers_of_loc
Summary:
We only want to emit constrained writes checks on unannotated variables in the new env. Otherwise, we get spurious error from code like this:
```
var x: number => string = (x) => 'str';

x = (x) => 'str'; // Error parameter of unknown type is incompatible with number
```

This happens because if we only do a constrained write check, and not a subtyping check, on the assignment then the parameter x in the assignment never gets an inferred type.

The first step towards avoiding emitting these errors is being able to distinguish between annotated and unannotated variables at the point of the assignment, which this diff does.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D33109946

fbshipit-source-id: 1317201788b3062506f6402ee7810f4e850c04ed