[name-def] Improved representation of dependency cycles
Summary:
The way that the name_def_ordering looked for cycles previously was broken--it assumed that any SCC had a cycle that visited every node exactly once, which of course is not how SCCs work. This changes the result to report separately the SCCs rather than looking for cycles within them, and also reports illegal self-cycles, values that depend on themselves.

This also now allows cycles between fully annotated values--in principle, this is no riskier than cycles between types.

Reviewed By: panagosg7

Differential Revision: D32325708

fbshipit-source-id: 224925a1dbbd4bfa5897193db481a9a46201068f