Config options for cycle errors and resolved environment
Summary: This creates some currently unused config options for use in moving forwards on the fully resolved environment. There's a new option for raising errors when cycles are found, and separately there's an option for actually creating the fully resolved environment--this is a new confguration of the existing `env_mode` flag. Finally, this renames the old statement-based reordering flag to try to reduce confusion with the new cycles work. This flag can be removed eventually, probably soon.

Reviewed By: panagosg7

Differential Revision: D33228390

fbshipit-source-id: a144041473e9bf7121671d6e8067d1d4a29d41b3