Move introduce_files into parsing phase
Summary:
Before this diff, after parsing, we would iterate over all parsed and unparsed
files, adding them to the "info heap" which stores the module provided by each
file.

However, we know the set of parsed/unparsed files during parsing. We can easily
compute the provided module and add this information to the heap during parsing.

This means we can skip introduce_files entirely. We don't need to read parsed
files' docblocks. We don't need to return unparsed files' docblocks from the
parse phase, only to pass them back into workers later on.

We can also share the same code for commit_modules between saved state and
non-saved state inits, which is a nice simplification.

Finally, it should be faster! We pay a little bit more in parsing to write the
info heap entries, but we needed to do that anyway.

Reviewed By: mroch

Differential Revision: D33307416

fbshipit-source-id: 7315ec8249cad17904f549ca29de5ab2304e84b6