Clean up package_json parsing
Summary:
Before this diff, when we saw a file named package.json, we would parse the
source file as JSON, reconstruct the AST into the form `module.exports = ...`,
then parse that using the Package_json module. Error handling was a bit
complicated, since JSON parsing can emit errors as well as Package_json parsing.

This diff uses a simpler method to parse package JSON files. Instead of parsing
arbitrary JSON, then parsing that again, we parse JSON objects directly. It
turns out that this subsumes the need for Package_json parse errors entirely.

This diff also includes a drive-by refactoring. In `fail` mode, parsing now
raises an exception with a non-empty list of errors. This simplifies handlers
which had to deal with the empty list case.

This diff also fixes two edge case bugs:

1. `do_parse` used to write to the package heaps, but we should only write to
the heap from `reducer`. `do_parse` is called from `check_contents`. In
practice, I don't think that we would call `check_contents` for a package_json
file, so this was likely benign.

2. When `fail` is `true` and `parse_json` failed, we would not register the
package.json file with the heap. If `fail` is `false`, we would register the
package.json file with the heap. The expected behavior here is unclear, but this
diff makes the outcome consistent -- we raise a parse error and add to the heap.

Reviewed By: mroch

Differential Revision: D33356336

fbshipit-source-id: c51e33db73b07e42a3ed725bcbf810866e28d833