First cut at cycle errors
Summary: This is an initial stab at errors for strongly connected components that need annotations in order to be typechecked, as well as errors for values that depend on themselves. These errors are currently not suitable for consumption by users, but they contain enough information that we can use them to understand existing codebases and what kinds of errors are likey to appear.

Reviewed By: panagosg7

Differential Revision: D33178257

fbshipit-source-id: ec6ca004629c816b5b947867b1ed9d18d51c9b13