Initial work on fully resolved environment
Summary:
This creates the initial implementation of the fully resolved environment. Currently, this only supports limited constructs: variable declarations (with or without annotations) and assignments, including most forms of destructuring assignments, and for loops (included only because they're modeled using the same tools as assignments and declarations)

This design consumes the name->def map produced in `name_def.ml`, which maps the location of a variable binding to the AST node that defines its type. We then typecheck those nodes using `env_resolution.ml`, which itself calls into `statement.ml` to actually perform typechecking. It does so in the same manner that statment.ml itself does. We then populate the new_env with the resulting type.

In addition, this diff includes some TODO work on caching the resulting typed AST nodes produced by typechecking during the creation of the env. We don't do anything with this now--that's a later diff.

When we populate the new_env with entries, we also explicitly mark them as resolved. This will prevent future calls to `set_env_entry` from adding new bounds to the types in the environment, and helps validate that we're doing the right thing. That said, currently this work is slightly mis-named: the types that we store in the environment are not actually necessarily `Resolved` tvars--they can still be open. Actually sealing them is future work.

Reviewed By: panagosg7

Differential Revision: D33228702

fbshipit-source-id: 05681ed89d90ca76d6f51f44b47f86c442f7288c