Move exports to the new heap
Summary:
Instead of a separate ExportsHeap, this diff stores the exports data as a field
in the "checked file" object accessible through CheckedFileHeap.

We load exports from saved state before we have any parse data like ASTs. Later,
`ensure_parsed` will fill in the parse data. When writing parse data, we first
check for an existing entry. If one exists, we update it with the new parse
data.

Reviewed By: mroch

Differential Revision: D33119724

fbshipit-source-id: c8dfbffab9d75075e9dc2ebffce117b9fb19496e