Unify info and checked file heaps
Summary:
The InfoHeap has entries for all parsed and unparsed files, while the
CheckedFileHeap only has entries for parsed files.

This diff unifies both into a new heap: FileHeap. The entry is a variant type,
either parsed or unparsed. The variant is encoded using the tag on the heap
object.

Reviewed By: mroch

Differential Revision: D33371795

fbshipit-source-id: 6a433f81d29ca53fba6254afe083ebf662cea38a