Remove Module_js.checked_file
Summary:
I always thought this function was a bit difficult to understand from its name.
It looks up a file's info (unsafely, assuming the file is tracked), and returns
the `checked` bit in the info record.

We can answer this question more simply by checking the tag on the heap object's
header. I also inlined this logic into the call sites, since I think it's easier
to understand when made a bit more explicit -- get file, see if checked.

This change is working toward removing the `info` record entirely.

Reviewed By: mroch

Differential Revision: D33386060

fbshipit-source-id: 052271a6ba2f9fdb1783175b5b9128a1d644c627