improve precision of annotation requirement on function assignment
Summary:
The way we determined whether an assignment `lhs = rhs;` had a contextual type
available for the typing of `rhs` was by checking to see if `lhs` was annotated. With
LTI's constrained-writes we can relax this requirement by assuming there is a type
available iff the assignment is not a provider for `lhs`.

This way, we won't require an annotation anymore for `x` in
```
let y = (x: number) => 3;
y = (x) => 3;
```
This approach is not yet complete for more complex patterns:
```
var n = (x: string) => 1;
var null_ = null;
[n, [null_]] = [
    (y) => 1,   // TODO should not need annotation
    [(z) => 1], // error needs annotation because null_ pattern is a provider
];
```
This pattern is considered a provider as a whole, as it is a provider for `null_`. So for now
we still require an annotation for `y`. This will effectively be resolved when we are able to
inspect the "type of the pattern" before we consider the RHS of the assignment.

Reviewed By: mvitousek

Differential Revision: D33271916

fbshipit-source-id: d2084048dd529a621b92d20a0018341560c7371a