Sentinel refinements
Summary:
Sentinel refinements now record the location of the value to which the member expression is compared. This allows us to look up that the type of that same expression later so that we can actually convert into a `LeftP (SentinelProp)`

You'll also notice some changes to `NullishCoalesce` handling in the name_resolver, which previously called `null_test`. This is no longer possible, since `null_test` now requires an explicit `other` argument, which is not present in a nullish coalesce. Instead, we manually apply a NotR MaybeR refinement. We do lose the ability to add sentinel refinements to the LHS of a nullish coalesce with this change, but that seems like a minor regression, especially considering Flow's current handling of nullish coalescing. We could restore that bheavior one day if we would like to, as I detail in the comment in `logical_refinement`.

`Passed: 369, Failed: 108, Skipped: 269, Errored: 0`

Reviewed By: panagosg7

Differential Revision: D33146283

fbshipit-source-id: d19dfc40b459cf673d9cbd0bdb4c65b0dddbb980