Fix refinements in default case of switch statement
Summary:
`conjunct_all_refinements_for_key` was an ok implementation for modeling the default refinement BEFORE heap refinements, but with heap refinements you can generate refinements for multiple keys, which would be ignored by `conjunct_all_refinements_for_key`. There's no reason that we need to conjunct all the refinements for only a single key here either-- these refinement scopes that we are conjuncting are generated strictly from the eq_test we perform between the switch discriminant and the case expressions.

This diff changes that function to just conjunct all the refinements in scope. So now, in cases like the `default_switch_refinement` test case added, instead of only trying to conjunct refinements on the heap key we also conjunct the refinements on the base of the heap key too.

Reviewed By: panagosg7

Differential Revision: D33164213

fbshipit-source-id: 4d35bfe88d736d1516f3938a5d34c33baa493bd6