Cannot reassign constant errors
Summary:
This diff emits errors in the name_resolver when a constant is reassigned. We use the stored binding kind to determine the type of the binding in `pattern_identifier` and then we use the passed in `kind` to pattern identifier to determine:
1. If this is a declaration or an assignment
2. If it is a declaration, what kind of declaration it is.

For consts, we want to error on both redeclarations AND reassignments.

`Passed: 343, Failed: 133, Skipped: 269, Errored: 0`

Reviewed By: mvitousek

Differential Revision: D33142376

fbshipit-source-id: 7ae25ffc553f9b7712e72b8e27bfb68116105708