Fix try-catch control-flow handling
Summary:
The name_resolver's implementation for merging environments for try/catch/finally was implemented very conservatively-- every time we enter a catch or finally block we were havocing every env entry.

Statement.ml is more precise. When we enter the catch block, we take the starting state of the env before the try block and merge it with the ending state after the try block. This represents the fully range of possibilities that could reach the catch block. Similarly for finally, we merge the start of the catch block with the end of the catch block.

Hilariously, there were comments in the name_resolver here detailing this exact strategy to be more precise as well. This diff implements that strategy.

It has no effect on test case numbers because the try-catch case has a constrained writes error that appears in the new-env but not the old one.

Reviewed By: panagosg7

Differential Revision: D33165684

fbshipit-source-id: cf9f7d8282a34915cf07a25e7edaac769bb97578