Handle delete
Summary:
[We model delete by writing `undefined` to the argument](https://www.internalfb.com/code/fbsource/[a6f889ef70e55afa33991cb20a1aadac9c65b831]/fbcode/flow/src/typing/statement.ml?lines=6146-6149). This diff adds that write to the name_resolver, too. Importantly, this write should also be counted as a def_loc by the provider analysis so that we can look up the providers of the argument to ensure that `undefined` is compatible with them.

We handle this by adding an `Undefined` Va.t, which gets translated into void. We did not re-use the type-equivalent `uninitialized` because the reasons for undefined/uninitialized should be different.

No test cases are fixed because the test that looks at delete has code like:
```
var x = 42;
delete x;
```

In the new environment, this errors because `undefined` is incompatible with number, but in the old environment this just widens x to `void | number`.

Reviewed By: mvitousek

Differential Revision: D33167762

fbshipit-source-id: 8af00588dd32b9c15445db28901042f2625414a4