Function reassignment errors
Summary:
This diff adds support for function reassignment errors to the new-env. We error if the original binding was a declared function, but not if it was a regular one. This diff does NOT handle the "true" semantics of multiple declared functions of the same name. That will come later in this stack.

`Passed: 347, Failed: 129, Skipped: 269, Errored: 0` no change, only partial changes to some tests

Reviewed By: mvitousek

Differential Revision: D33144537

fbshipit-source-id: 047387919518243d21841d60826bd39bca3f2a05