Do not process assignments on reassignment/redeclaration errors
Summary:
When we see a reassignment/redeclaration error we shouldn't also emit a subtyping check against the providers of the binding. For example:

```
const x: string = 'str';
x = 3; // Cannot reassign x, but not an 3 is incompatible with string error
```

Since subtyping checks are not emitted by the name resolver, we record that a write that causes an assignment error as "NonAssigning", which tells the new-env to not perform the subtyping check against the providers.

Note that there are some cases where this does not model the runtime as precisely as possible. For example, this code causes a redeclaration error:

```
class A {}
let A = 3;
```

But we will continue to treat `A` as a class in the rest of the program. This is a deliberate design choice, but I am open to discussing this and changing the semantics on case-by-case basis.

`Passed: 346, Failed: 130, Skipped: 269, Errored: 0`

Reviewed By: mvitousek

Differential Revision: D33142902

fbshipit-source-id: 8d52d86ed059f27f70271a4451c1d6d8177bba7a