Fix latent crash with redeclared bindings
Summary:
Since re-declared bindings don't count, we need to make sure that code that deals with redeclared vars does not actually trigger a subtyping check. Otherwise, we end up looking up a write that does not exist according to the name_resolver.

See the added test case for an example of a crash that is fixed by this change.

Reviewed By: panagosg7

Differential Revision: D33144770

fbshipit-source-id: f8cdd83a72b5543aa6b17c04addbcae8a621488a