use before declaration errors for classes
Summary:
The previous diff added use-before-declaration errors for all kinds of bindings _except_ for classes. This diff adds those errors for classes.

We use `UndeclaredClass` (renamed from `UninitializedClass`) to represent the write state of classes in a scope before the class binding has been seen. We separate this from the already-existing `Undeclared` because using the class name at the type-level is allowed even if the binding has not yet been encountered:

```
let x: ?A = null; // No error
x = new A(); // Error, use-before-declaration
class A {}
x = new A();
```
If we encounter an UndeclaredClass when for_type is false then we emit an error the same way we do with an Undeclared env val.

No effect on test failures, but instances where we were missing errors for use-before-declared classes now report errors.

Reviewed By: mvitousek

Differential Revision: D33259693

fbshipit-source-id: f0000e2d5c34ab580af8ec378774d325b0821e2b