remove coverage data for deleted files
Summary:
we store coverage for all checked files in `env.coverage`. we never clear it out, only update it during check, in `Types_js.update_check_results`.

but we can clear out unparsed and deleted files to save space/avoid leaks. and by doing so, we don't need to check whether the files in `env.coverage` are in `env.checked_files`. i'm trying to see if we can reduce or eliminate `checked_files`.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D33538977

fbshipit-source-id: 5a4ed33af72cfa2161aa577f237c667a44230c33