fix watchman when including missing paths on Windows
Summary:
when there are multiple included paths, we call `watchman watch-project` on each of them, and watchman tells us the project root and relative path. then we subscribe to the root, watching all the relative paths we care about.

when an included path doesn't exist, we relativize it ourselves, and subscribe to that anyway, so that we're notified if it is created.

the way we were doing that was broken on Windows. watchman normalizes the paths to use `/`, but we were using `\`.

Changelog: [fix] Fix Watchman file watcher on Windows when an include path doesn't exist

Reviewed By: samwgoldman

Differential Revision: D33829886

fbshipit-source-id: ee3f1b737811226e3ac0b0618c85c1af32c61eae