Mark all modules provided by old files as dirty
Summary:
When finding dirty modules which need to be committed, we find both "old" and
"new" modules. A changed or deleted file provided a file module and maybe a
haste module. These are "old" modules and are possibly affected by the change.

New or changed files provide modules as well. These modules are "new" modules.
Taking old and new modules together gives us the set of dirty modules which need
to be re-committed.

Before this change, when calculating old modules, we selected only modules which
were actually provided by the changed or deleted file. For example, if the file
Foo.js was a duplicate provider the haste module Bar -- i.e., it was not
chosen to be the provider by the previous call to commit_modules -- we would
exclude Bar from old modules.

Excluding these modules is valid, because we will never re-pick a different
provider for such a module. I know of no bug caused by this behavior, but
removing it lets us make things simpler.

After this change, we include all modules provided by a changed or deleted file,
even if that file is not the chosen provider.

This change prepares for subsequent changes:

1. I will simplify the maintenance of duplicate provider errors. Instead of
storing these errors keyed by file, we will store them keyed by module name, and
invalidate all dirty modules. Without this change, we would fail remove
duplicate provider errors when the duplicate provider is deleted, because the
module is not considered dirty.

2. I will move the "all providers" data into shared memory and update it in
parallel in workers. Note that we update the all providers hashtable using all
old modules. Marking them all as dirty means we can update all providers and
choose a single provider at the same time in workers.

Reviewed By: mroch

Differential Revision: D33837427

fbshipit-source-id: 109eb571974c4ca1ab23a75ee6e59eae07373f00