make signature-verification-failure an InferError
Summary:
`signature-verification-failure` is a weird lint. it errors even when lints are set to `all=off`, and `flowlint signature-verification-failure:off` doesn't suppress it even if you set `all=error`.

AFAICT, it's only a lint out of convenience.

Changelog: [misc] Change `signature-verification-failure` from a lint to a normal error. It is no longer valid in the `[lints]` section of `.flowconfig`.

Reviewed By: panagosg7

Differential Revision: D30818890

fbshipit-source-id: 8162f84b0c514d84a9c4edb874890c8795a879cc