better error message for [invalid-exported-annotation]
Summary:
In many cases, when merge fails with [invalid-exported-annotation] it is because
an export is trying to operate on an `$ObjMap(i)` type. This diff improves in the error
messages we produce in these cases by suggesting the use of `$ObjMapConst`
and `$KeyMirror`, which correspond to commonly used cases of these utility types
and are supported by annotation inference.

Changelog: [internal]

Reviewed By: bradzacher

Differential Revision: D33997026

fbshipit-source-id: 60ccde3d725a3278bed7212e0fa3133657d33952