Get rid of synthetic default case
Summary:
One of the more confusing parts of our switch handling code is how we handle switches without defaults. We currently add in an extra "synthetic" default case when one is not present and then pass it through the loop as normal. However, this can cause extremely confusing behavior that takes a deep understanding of this code to catch.

For example, an empty synthetic default case will always fall through. This subtle behavior was super confusing when debugging the case from the previous diff:
```
switch (x) {
  case 'bar':
    x = 'foo';
    break;
}
(x: 'foo'); // Error ??
```

The current implementation is convenient in that some code duplication can be avoided, but it would be better if we handle the absence of a default explicitly and separately so that this code is easier to understand, as this diff does.

Reviewed By: panagosg7

Differential Revision: D33956399

fbshipit-source-id: 271a4cfd28ba04f29c9ad4a9335784101fcf4c1e