Return addr from hh_add
Summary:
When two threads race to add to the shared hash table, the first one to write
wins. This diff makes it possible to observe the winning value in both threads.
The successful thread will see that the return value matches the addr they tried
to add, the failing thread will see the return value differ.

Historically, races to write in Hack/Flow have not needed this information due
to an invariant: threads racing like this must be trying to write "equivalent"
values.

Soon, entities will bring reference equality to shared memory, meaning it is not
possible for two threads to write an equivalent entity. This change makes it
possible to build parallel algorithms that can cope with this.

Reviewed By: mroch

Differential Revision: D33746703

fbshipit-source-id: 14c75e4b82474fbdd03d17b4a5583b729d4d186a