Re-organize shared heap tag values
Summary:
We store different kinds of objects in shared memories. Some need to be scanned
for pointers, some shouldn't, some are serialized and compressed, and each needs
to be read a bit differently.

We use ranges of tag values to identify these cases. This diff flips that range
so that scanned objects come first instead of last. I think this makes the logic
a bit simpler, but ultimately it's equivalent.

This diff also lifts the tag definitions higher in sharedMem.ml so it can be
used in the definition of NoCache and NoCacheTag. This is a real improvement,
because we can avoid some magic numbers in those definitions which were ticking
time bombs.

Reviewed By: mroch

Differential Revision: D33746702

fbshipit-source-id: a0b280731722262b332658e452b1829872003832