Entities
Summary:
This diff extends shared memory with a new kind of data: entities. Entities are
a primitive that can be used to build partially-persistent data structures,
specifically having at most two versions at a time.

This kind of persistence is used during rechecks, where we keep a consistent
snapshot of the type state while writing the new type state. This has
transactional behavior, meaning that readers can observe the consistent snapshot
while the mutator is updating. Committing the transaction "publishes" the new
data to subsequent readers.

Before entities, we implemented these transactions using oldify/revive. The
server would keep track of oldified keys and dispatch readers to the appropriate
key.

Entities use a different approach. There is a global transaction counter and
each entity also has a version derived from the transaction counter. The entity
has two slots for data, containing either (a) the committed data and null or (b)
the committed data and latest/in-progress data.

Readers and mutators can load the appropriate data by comparing the global
transaction and the entity version. This does not require additional
bookkeeping of the oldified set.

See the next diff in this stack for an example of replacing oldify/revive with
entities.

Reviewed By: mroch

Differential Revision: D33746701

fbshipit-source-id: 3eb5db1d3fefe8d231df1a84925901b09e122056