[glean] Show version of `all` schema supported by the indexer.
Summary:
We want indexers to pin the version of the `all` schema that was
supported when building the database. For indexers that produce
json we are doing this by asking the indexer for the version it
supports such that the json ingesting procedure can use it to
fill the relevant database property like in D34024281.

This diff adds to the `--schema-version` flag to the indexer,
which will output the supported version of the `all` schema.
This will need to be changed when the indexer is modified to
produce a newer schema version

Reviewed By: nhawkes

Differential Revision: D34073376

fbshipit-source-id: ed947a379caafe94781671a1e918276b8eb3edab