Avars are not needed for require tvars
Summary:
This diff partially reverts D28496027 (https://github.com/facebook/flow/commit/628f1c6b201279bfc44cb6e0940516e8024a84dc), which abstracted the creation of require
tvars. Annotation constraints ensure that types are resolved in a single step,
and we use avars to implement that behavior.

Signatures are converted to types for dependencies of the checked file, but the
checked file itself still uses tvars, including unresolved tvars, for inference.
We call add_require_tvars on checked files only, not dependencies, so we don't
need to use avars.

To avoid accidentally using tvars in annotation constraints, the Flow_js module
is shadowed in the Check_service module. This restriction is not strictly
necessary in the part of this file that sets up the checked file's context. As a
future improvement, we might consider factoring the dependency context code --
`dep_module_t` and related functions -- into its own module.

This change is not a bug fix. Creating avars here does not cause any bad
behavior. In later changes, it will be useful to know that all avars in a
context correspond to unresolved annotation constraints.

Reviewed By: panagosg7

Differential Revision: D34109928

fbshipit-source-id: c369fe8cc20fce2d6c63831d160c7d83f7091a3f