Fix parsing of f<<T>()=>S>()
Summary:
As reported [here](https://github.com/facebook/flow/issues/8824), we parse function calls with explicit instantiation incorrectly if the type begins with `<`. In such cases, we lex `<<` as the left shift operator. We don't have similar problems with `>>` or with `<<` in type definitions because we lex those cases as types, but we lex function call explicit instantiation as regular tokens. This diff switches to parsing as types eariler.

Closes #8824

Changelog: [parser] Fixed a bug where type parameter instantiations that start with `<` (e.g., `f<<T>()=>S>()`) would be incorrectly parsed as the left shift operator.

Reviewed By: mroch

Differential Revision: D34050306

fbshipit-source-id: 0c16b393e1f8e7c0fdc8c9439d97256379e4eb66