Don't flow specific to general when refining
Summary:
Refinements don't really need to affect the actual general type of a variable in the old env. With this change, we no longer flow specific to general when we refine. This work was motivated by the desire to produce better types in codemods--in particular, without it, we can have `React.Node` be inappropriately inlined as a useless object type.

With this change, we see semantic changes in a few places. Since refinements no longer affect general tvars, if a tvar is only meaningfully constrained by its refinements, then we no longer will get errors when those refinement types are used. This comes into play in the `flow_to_general` test case added in the previous diff, where a variable has `any` and `number` as lower bounds, `number` only due to a refinement. We now no longer get errors when we check this type against empty. This seems like a more consistent behavior, given that the only lower bound that comes from a value is `empty`.

We also lose errors in some cases for literal subtyping checks, but these errors are apparently universally duplicates, so that's fine.

In order to maintain errors on inappropriate `instanceof` checks, we add a new check to `refine_with_predicates` that adds a post inference check that the refined value is a subtype of the general type.

Finally, we get a large set of new legit errors in cases where the refinement to general flow caused a type to inappropriately widen. This allowed us to miss cases where string unions were tested against strings not in the union, which allowed misspellings to be uncaught!

Changelog: [errors] Modified the behavior of `instanceof` checks and other refinements to more consistently detect when impossible or invalid refinements are made. This will result in new errors in such cases.

Reviewed By: panagosg7

Differential Revision: D34320070

fbshipit-source-id: fba14c852d3da589731d36e14a0c24a9497f14c5