Additional simplifications for Fbt
Summary:
The `Fbt` type contains string, and in codemods its been common for us to synthesize types like `Fbt | string`. This adds a hardcoded fix that filters out `string` from unions that contain `Fbt`.

This work restructures the union cases in the hardcoded fixer, in order to allow multiple fixes to be applied to the same type in a single pass.

Reviewed By: panagosg7

Differential Revision: D34353405

fbshipit-source-id: 2629d9593729d74fcd8296329be095a73a3c1725