[ez] Always filter deep empties in annotate-declarations codemod
Summary:
The annotate declarations codemod filters out types inferred to contain empty, which is necessary because these empties can pollute type significantly due to the post-inference nature of the check. However, this ability was hidden under a flag. It's hard to imagine cases where we wouldn't want this check, so let's make it the default.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34352218

fbshipit-source-id: ecbd1d303bdbcf14fd1436a88a7f067810fe9253