Generalize unions containing null or void to maybes
Summary: Perhaps a controversial change? This changes type synthesis of inferred unions that contain either `null` or `void` to be considered to have both of them, so we can represent them as `?T` maybe types. This is of course not semantically preserving, but the risk of widening to maybe types seems low.

Reviewed By: panagosg7

Differential Revision: D34353404

fbshipit-source-id: b67e0771eeba0cbd0afc68765e5ec1795f85d539