Recheck dependents when declaration file becomes initial provider
Summary:
It is possible that a require will not resolve to any file. In this case, we
emit an error that the module can not be resolved.

However, the file might be introduced later, so we also need to store some
information to track these dependencies. We do this during "resolve requires" by
storing file paths as "phantom dependents" (sic: should be dependencies).

Later, in calc_direct_dependents, we scan every file's phantom dependents. If a
new, changed, or deleted file is in a given file's phantom dependents set, then
that file needs to be rechecked.

Before this change, we would not find .flow files in the phantom dependents set,
because we never added those paths. Observe that Module_js.path_if_exists checks
for both .ext and .ext.flow files, but we only record the .ext paths.

Storing .ext and .ext.flow in phantom dependents would use more space. Instead,
this diff takes a different approach. In calc_direct_dependents, we normalize
the changed/new/deleted file names before comparing to phantom dependents.

(This normalization approach could also be applied for file_exts, which come
from the module.file_ext configuration. Currently we store a path for each file
ext which is wasteful.)

Reviewed By: panagosg7

Differential Revision: D34452977

fbshipit-source-id: 4ce82cf74ed01554acfd34bbe98bbe9500e9cc98