Update wording of "missing property" reason
Summary:
`RMissingProperty` is used when accessing a non-existent property from an exact object (non-exact objects produce `RUnknownProperty`).

I update the wording here to make it more clear. While this isn't used much currently, in the future some of my changes may make this more common so it is better to fix this up.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34398605

fbshipit-source-id: b3d568cba57d8f4ba2b5c2fff970dbece495dfc8