Add capability to mark new-env tests as passing
Summary:
As prior diffs have mentioned, many fixes to the new-env are no longer causing tests to pass due to superficial reasons. This diff adds the ability for a test to include an `test.exp.new_env` file which records the results expected in the new_env. When `runtests` is run with `-e` and a `.exp.new_env` file is present, it uses that file to compare the results instead of the results in the `.exp` file.

I took this approach instead of making a list of "passing" tests to not run so that I could make sure that I do not regress any tests that I consider to be passing.

Reviewed By: mvitousek

Differential Revision: D33168918

fbshipit-source-id: 665c8335f287666790532aa776d1ff6d94b7dd7b