check literal subtypes
Summary:
When we see a refinement like `x === 'a'` we check to make sure that `'a'` is a subtype of `x`'s general type in a post-inference pass. Since we don't have access to types in the name_resolver, we can't do exactly that. Instead, we record the locations of the refiner and the binding so that we can emit the check in a post-inference pass.

This diff also fixes a minor bug with the `ub_visitor` in that post-inference check. The parent class is the resolver_visitor, but we might still pass in an OpenT to the ub_visitor, in which case we would just merge the tvar and never traverse the resulting type. The change here calls the parent `type_` before attempting to substitute a GenericT for its upper bound.

`Passed: 382, Failed: 86, Skipped: 287, Errored: 0`

Reviewed By: panagosg7

Differential Revision: D33168772

fbshipit-source-id: 9c701342757d43f6ce3acf0d5c10e5e4362daeb6