Add binding_kind to provider analysis entries
Summary:
We are going to use the provider_analysis to report every declared function binding as a provider for that binding's type. When we want to find the type of a declared function we will query the provider reported types and then intersect them together. In order to do that, we need to keep track of the binding kind of a provider so that we can implement the special provider rules for declared functions.

Changelog: [Internal]

Reviewed By: panagosg7

Differential Revision: D34399110

fbshipit-source-id: 34c85e692dbd0b1a59b2ff37d122c5a8fa16769c