Report declared functions as providers
Summary:
The intersection of the declared functions are the providers ofa  declared function. The signature of the implementation function is not taken into account when computing the declared function's type, as Flow operates today.

This diffs adds this behavior to the provider analysis (see added test case), and the next diff has the new-env actually leverage this new behavior to type declared functions.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D34399343

fbshipit-source-id: 80e955b117ed9141a1bbb02de71218875fb02833